KRACOEUS-6956 : repack propdev move ProposalType.java to common
