+++ coeus-code/src/main/java/org/kuali/coeus/common/framework/type/ProposalType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.proposaldevelopment.bo;
+package org.kuali.coeus.common.framework.type;
 
 import org.kuali.coeus.sys.framework.model.KcPersistableBusinessObjectBase;
