+++ coeus-code/src/main/java/org/kuali/kra/negotiations/printing/xmlstream/NegotiationActivityXmlStream.java
@@ -17,12 +17,12 @@ package org.kuali.kra.negotiations.printing.xmlstream;
 
 import org.apache.xmlbeans.XmlObject;
 import org.kuali.coeus.common.framework.print.stream.xml.XmlStream;
+import org.kuali.coeus.common.framework.type.ProposalType;
 import org.kuali.coeus.sys.framework.model.KcPersistableBusinessObjectBase;
 import org.kuali.kra.negotiations.bo.*;
 import org.kuali.kra.negotiations.printing.NegotiationActivityPrintType;
 import org.kuali.kra.printing.schema.*;
 import org.kuali.kra.printing.schema.NegotiationsDocument.Negotiations;
-import org.kuali.kra.proposaldevelopment.bo.ProposalType;
 import org.kuali.rice.core.api.datetime.DateTimeService;
 import org.kuali.rice.krad.service.BusinessObjectService;
 import org.kuali.rice.krad.service.DocumentService;
