+++ coeus-code/src/main/java/org/kuali/kra/proposaldevelopment/bo/LookupableDevelopmentProposal.java
@@ -17,6 +17,7 @@ package org.kuali.kra.proposaldevelopment.bo;
 
 import org.apache.commons.lang3.StringUtils;
 import org.kuali.coeus.common.framework.sponsor.Sponsor;
+import org.kuali.coeus.common.framework.type.ProposalType;
 import org.kuali.coeus.common.framework.unit.Unit;
 import org.kuali.coeus.sys.framework.model.KcPersistableBusinessObjectBase;
 import org.kuali.coeus.sys.framework.service.KcServiceLocator;
