+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/DevelopmentProposal.xml
@@ -340,11 +340,11 @@
                 </property>
 	        <property name="required" value="true" />
         	<property name="control">
-                        <bean parent="SelectControlDefinition" p:businessObjectClass="org.kuali.kra.proposaldevelopment.bo.ProposalType" p:valuesFinderClass="org.kuali.coeus.sys.framework.keyvalue.ExtendedPersistableBusinessObjectValuesFinder" p:includeKeyInLabel="false" p:keyAttribute="proposalTypeCode" p:labelAttribute="description"/>
+                        <bean parent="SelectControlDefinition" p:businessObjectClass="org.kuali.coeus.common.framework.type.ProposalType" p:valuesFinderClass="org.kuali.coeus.sys.framework.keyvalue.ExtendedPersistableBusinessObjectValuesFinder" p:includeKeyInLabel="false" p:keyAttribute="proposalTypeCode" p:labelAttribute="description"/>
                 </property>
                 <property name="optionsFinder">
                         <bean class="org.kuali.coeus.sys.framework.keyvalue.ExtendedPersistableBusinessObjectValuesFinder">
-                                <property name="businessObjectClass" value="org.kuali.kra.proposaldevelopment.bo.ProposalType"/>
+                                <property name="businessObjectClass" value="org.kuali.coeus.common.framework.type.ProposalType"/>
                                 <property name="includeKeyInDescription" value="false"/>
                                 <property name="keyAttributeName" value="proposalTypeCode"/>
                                 <property name="labelAttributeName" value="description"/>
