+++ coeus-code/src/main/resources/repository.xml
@@ -330,7 +330,7 @@
             <foreignkey field-ref="moduleCode" />
         </reference-descriptor>
     </class-descriptor>
-	<class-descriptor class="org.kuali.kra.proposaldevelopment.bo.ProposalType" table="PROPOSAL_TYPE">
+	<class-descriptor class="org.kuali.coeus.common.framework.type.ProposalType" table="PROPOSAL_TYPE">
 		<field-descriptor name="proposalTypeCode" column="PROPOSAL_TYPE_CODE" jdbc-type="VARCHAR" primarykey="true" />
 		<field-descriptor name="description" column="DESCRIPTION" jdbc-type="VARCHAR" />
 		<field-descriptor name="updateTimestamp" column="UPDATE_TIMESTAMP" jdbc-type="TIMESTAMP" />
